Emotion_grpo/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "emotion-grpo"
version = "0.1.0"
description = "A standalone VERL-based GRPO research scaffold for intrinsic reward experiments."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "numpy>=1.26.0",
  "pandas>=2.2.0",
  "pyarrow>=16.0.0",
  "PyYAML>=6.0.0",
  "transformers>=4.45.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0.0",
]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
